requirements.txt
flask>=3.0
PyMuPDF>=1.24
Pillow>=10.0
gunicorn>=21.2
